backend/requirements.txt
torch
diffusers
transformers
numpy
pillow
flask
flask-cors
accelerate
xformers
python-dotenv
requests_toolbelt
imageio-ffmpeg
opencv-python
openai